refactor(admin): rename inbox shortcuts to admin triage shortcuts
Prepares src/lib/components/admin/inboxShortcuts.ts for reuse in the
detail view's queue mode by giving it a name that no longer implies
it belongs only to the inbox page. Export names are unchanged.
